feat: Added sugarcane Top feat: Added ClymeStatistics

+++ src/main/java/solutions/misi/clymeskyblockcore/utils/SortingUtils.java
@@ -6,12 +6,12 @@ import java.util.*;
 
 public class SortingUtils {
 
-    public static HashMap<OfflinePlayer, Integer> sortByValue(HashMap<OfflinePlayer, Integer> hm) {
-        List<Map.Entry<OfflinePlayer, Integer>> list = new LinkedList<>(hm.entrySet());
+    public static HashMap<OfflinePlayer, Long> sortByValue(HashMap<OfflinePlayer, Long> hm) {
+        List<Map.Entry<OfflinePlayer, Long>> list = new LinkedList<>(hm.entrySet());
         list.sort(Map.Entry.comparingByValue());
         Collections.reverse(list);
-        HashMap<OfflinePlayer, Integer> temp = new LinkedHashMap<>();
-        for(Map.Entry<OfflinePlayer, Integer> aa : list) temp.put(aa.getKey(), aa.getValue());
+        HashMap<OfflinePlayer, Long> temp = new LinkedHashMap<>();
+        for(Map.Entry<OfflinePlayer, Long> aa : list) temp.put(aa.getKey(), aa.getValue());
         return temp;
     }
 }
